IAT-49:  Fixed build and test failures related to the logger moving from utility to service.
